finished messageForm with clickEvent
